% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{as.data.frame,Reader-method}
\alias{as.data.frame,Reader-method}
\title{Convert the Reader into a data.frame}
\usage{
\S4method{as.data.frame}{Reader}(x, row.names = NULL, optional = FALSE, ...)
}
\description{
Convert the Reader into a data.frame
}
