% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{initialize,Reader-method}
\alias{initialize,Reader-method}
\title{Create a new Reader}
\usage{
\S4method{initialize}{Reader}(.Object, filename, parser = "")
}
\arguments{
\item{.Object}{base object}

\item{filename}{path to the file to be parsed}

\item{parser}{name of the parser to be used; if not specified, auto-detected}
}
\value{
Reader wrapping the opened file
}
\description{
Create a new Reader
}
